Move GitCommandHelpers under Git/Commands
