+++ GitUI/CommandsDialogs/FormCleanupRepository.cs
@@ -3,6 +3,7 @@ using System.IO;
 using System.Linq;
 using System.Windows.Forms;
 using GitCommands;
+using GitCommands.Git.Commands;
 using GitCommands.Utils;
 using GitUI.HelperDialogs;
 using JetBrains.Annotations;
