+++ GitUI/CommandsDialogs/FormCreateTag.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Drawing;
 using System.Windows.Forms;
-using GitCommands;
 using GitCommands.Git.Commands;
 using GitCommands.Git.Extensions;
 using GitCommands.Git.Tag;
