+++ GitUI/CommandsDialogs/FormRevertCommit.cs
@@ -2,6 +2,7 @@
 using System.Drawing;
 using System.Windows.Forms;
 using GitCommands;
+using GitCommands.Git.Commands;
 using GitUI.HelperDialogs;
 using ResourceManager;
