+++ GitUI/HelperDialogs/FormResetCurrentBranch.cs
@@ -2,6 +2,7 @@
 using System.Diagnostics;
 using System.Windows.Forms;
 using GitCommands;
+using GitCommands.Git.Commands;
 using GitExtUtils.GitUI.Theming;
 using ResourceManager;
