+++ UnitTests/CommonTestUtils/GitModuleTestHelper.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Text;
 using GitCommands;
 using GitCommands.Config;
+using GitCommands.Git.Commands;
 
 namespace CommonTestUtils
 {
